Moved autoagents out of challenge folder
