+++ srunner/autoagents/agent_wrapper.py
@@ -17,7 +17,7 @@ import carla
 
 from srunner.challenge.envs.scene_layout_sensors import SceneLayoutReader, ObjectFinder
 from srunner.challenge.envs.sensor_interface import CallBack, CANBusSensor, HDMapReader
-from srunner.challenge.autoagents.autonomous_agent import Track
+from srunner.autoagents.autonomous_agent import Track
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
 
 MAX_ALLOWED_RADIUS_SENSOR = 5.0
