+++ srunner/autoagents/npc_agent.py
@@ -12,7 +12,7 @@ from __future__ import print_function
 import carla
 from agents.navigation.basic_agent import BasicAgent
 
-from srunner.challenge.autoagents.autonomous_agent import AutonomousAgent, Track
+from srunner.autoagents.autonomous_agent import AutonomousAgent, Track
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
